requirements.txt
streamlit
langchain
sentence-transformers
transformers
faiss-cpu
python-dotenv
huggingface_hub